feat(suite-native): warning for coin deletion
